requirements.txt
flask==3.0.3
fpdf2==2.7.9
